create base context
