+++ deepchecks/nlp/context.py
@@ -15,13 +15,11 @@ from operator import itemgetter
 
 import numpy as np
 
-from deepchecks.core import CheckFailure, CheckResult, DatasetKind
+from deepchecks.core.context import BaseContext
 from deepchecks.core.errors import (DatasetValidationError, DeepchecksNotSupportedError, DeepchecksValueError,
                                     ModelValidationError, ValidationError)
-from deepchecks.nlp.metric_utils.scorers import init_validate_scorers
 from deepchecks.nlp.task_type import TaskType
 from deepchecks.nlp.text_data import TextData
-from deepchecks.tabular.utils.task_type import TaskType as TabularTaskType
 
 __all__ = [
     'Context',
@@ -29,7 +27,6 @@ __all__ = [
     'TTextProba'
 ]
 
-from deepchecks.tabular.metric_utils import DeepcheckScorer, get_default_scorers
 from deepchecks.tabular.utils.validation import ensure_predictions_proba, ensure_predictions_shape
 from deepchecks.utils.typing import BasicModel
 
@@ -146,7 +143,6 @@ class _DummyModel(BasicModel):
     @staticmethod
     def _validate_prediction(dataset: TextData, prediction: TTextPred):
         """Validate prediction for given dataset."""
-
         if not isinstance(prediction, collections.abc.Sequence):
             ValidationError(f'Check requires predictions for {dataset.name} to be a sequence')
         if len(prediction) != dataset.n_samples:
@@ -245,7 +241,7 @@ class _DummyModel(BasicModel):
                                           f'dataset to be probabilities and sum to 1 for each row')
 
 
-class Context:
+class Context(BaseContext):
     """Contains all the data + properties the user has passed to a check/suite, and validates it seamlessly.
 
     Parameters
@@ -306,30 +302,16 @@ class Context:
         self._with_display = with_display
 
     @property
-    def with_display(self) -> bool:
-        """Return the with_display flag."""
-        return self._with_display
-
-    @property
-    def train(self) -> TextData:
-        """Return train_dataset if exists, otherwise raise error."""
-        if self._train is None:
-            raise DeepchecksNotSupportedError('Check is irrelevant for Datasets without train_dataset dataset')
-        return self._train
-
-    @property
-    def test(self) -> TextData:
-        """Return test_dataset if exists, otherwise raise error."""
-        if self._test is None:
-            raise DeepchecksNotSupportedError('Check is irrelevant for Datasets without test_dataset dataset')
-        return self._test
-
-    @property
     def model(self) -> _DummyModel:
         """Return model if exists, otherwise raise error."""
         if self._model is None:
             raise DeepchecksNotSupportedError('Check is irrelevant without providing predictions')
         return self._model
+
+    @property
+    def model_name(self) -> str:
+        """Return the name of the model."""
+        return 'Pre-computed predictions'
 
     @property
     def task_type(self) -> TaskType:
@@ -357,82 +339,3 @@ class Context:
                 f"but received model of type '{self.task_type.value.lower()}'"  # pylint: disable=inconsistent-quotes
             )
         return True
-
-    def get_data_by_kind(self, kind: DatasetKind):
-        """Return the relevant Dataset by given kind."""
-        if kind == DatasetKind.TRAIN:
-            return self.train
-        elif kind == DatasetKind.TEST:
-            return self.test
-        else:
-            raise DeepchecksValueError(f'Unexpected dataset kind {kind}')
-
-    def finalize_check_result(self, check_result, check, kind: DatasetKind = None):
-        """Run final processing on a check result which includes validation, conditions processing and sampling\
-        footnote."""
-        # Validate the check result type
-        if isinstance(check_result, CheckFailure):
-            return
-        if not isinstance(check_result, CheckResult):
-            raise DeepchecksValueError(f'Check {check.name()} expected to return CheckResult but got: '
-                                       + type(check_result).__name__)
-
-        # Set reference between the check result and check
-        check_result.check = check
-        # Calculate conditions results
-        check_result.process_conditions()
-        # Add sampling footnote if needed
-        if hasattr(check, 'n_samples'):
-            n_samples = getattr(check, 'n_samples')
-            message = ''
-            if kind:
-                dataset = self.get_data_by_kind(kind)
-                if dataset.is_sampled(n_samples):
-                    message = f'Data is sampled from the original dataset, running on ' \
-                              f'{dataset.len_when_sampled(n_samples)} samples out of {len(dataset)}.'
-            else:
-                if self._train is not None and self._train.is_sampled(n_samples):
-                    message += f'Running on {self._train.len_when_sampled(n_samples)} <b>train_dataset</b> data ' \
-                               f'samples out of {len(self._train)}.'
-                if self._test is not None and self._test.is_sampled(n_samples):
-                    if message:
-                        message += ' '
-                    message += f'Running on {self._test.len_when_sampled(n_samples)} <b>test_dataset</b> data ' \
-                               f'samples out of {len(self._test)}.'
-
-            if message:
-                message = ('<p style="font-size:0.9em;line-height:1;"><i>'
-                           f'Note - data sampling: {message} Sample size can be controlled with the "n_samples" '
-                           'parameter.</i></p>')
-                check_result.display.append(message)
-
-    def get_scorers(self,
-                    scorers: t.Union[t.Mapping[str, t.Union[str, t.Callable]], t.List[str]] = None,
-                    use_avg_defaults=True) -> t.List[DeepcheckScorer]:
-        """Return initialized & validated scorers in a given priority.
-
-        If receive `scorers` use them,
-        Else if user defined global scorers use them,
-        Else use default scorers.
-
-        Parameters
-        ----------
-        scorers : Union[List[str], Dict[str, Union[str, Callable]]], default: None
-            List of scorers to use. If None, use default scorers.
-            Scorers can be supplied as a list of scorer names or as a dictionary of names and functions.
-        use_avg_defaults : bool, default True
-            If no scorers were provided, for classification, determines whether to use default scorers that return
-            an averaged metric, or default scorers that return a metric per class.
-        Returns
-        -------
-        List[DeepcheckScorer]
-            A list of initialized & validated scorers.
-        """
-        if self.task_type == TaskType.TEXT_CLASSIFICATION:
-            scorers = scorers or get_default_scorers(TabularTaskType.MULTICLASS, use_avg_defaults)
-        elif self.task_type == TaskType.TOKEN_CLASSIFICATION:
-            scorers = []  # TODO: Complete that
-        else:
-            raise DeepchecksValueError(f'Task type must be either {TaskType.TEXT_CLASSIFICATION} or '
-                                       f'{TaskType.TOKEN_CLASSIFICATION} but received {self.task_type}')
-        return init_validate_scorers(scorers)
