+++ deepchecks/tabular/context.py
@@ -14,7 +14,7 @@ import typing as t
 import numpy as np
 import pandas as pd
 
-from deepchecks.core import CheckFailure, CheckResult, DatasetKind
+from deepchecks.core.context import BaseContext
 from deepchecks.core.errors import (DatasetValidationError, DeepchecksNotSupportedError, DeepchecksValueError,
                                     ModelValidationError)
 from deepchecks.tabular._shared_docs import docstrings
@@ -141,7 +141,7 @@ class _DummyModel:
 
 
 @docstrings
-class Context:
+class Context(BaseContext):
     """Contains all the data + properties the user has passed to a check/suite, and validates it seamlessly.
 
     Parameters
@@ -222,25 +222,6 @@ class Context:
 
     # Properties
     # Validations note: We know train & test fit each other so all validations can be run only on train
-
-    @property
-    def with_display(self) -> bool:
-        """Return the with_display flag."""
-        return self._with_display
-
-    @property
-    def train(self) -> Dataset:
-        """Return train if exists, otherwise raise error."""
-        if self._train is None:
-            raise DeepchecksNotSupportedError('Check is irrelevant for Datasets without train dataset')
-        return self._train
-
-    @property
-    def test(self) -> Dataset:
-        """Return test if exists, otherwise raise error."""
-        if self._test is None:
-            raise DeepchecksNotSupportedError('Check is irrelevant for Datasets without test dataset')
-        return self._test
 
     @property
     def model(self) -> BasicModel:
@@ -395,51 +376,3 @@ class Context:
         scorer_name = next(iter(scorers))
         single_scorer_dict = {scorer_name: scorers[scorer_name]}
         return init_validate_scorers(single_scorer_dict, self.model, self.train)[0]
-
-    def get_data_by_kind(self, kind: DatasetKind):
-        """Return the relevant Dataset by given kind."""
-        if kind == DatasetKind.TRAIN:
-            return self.train
-        elif kind == DatasetKind.TEST:
-            return self.test
-        else:
-            raise DeepchecksValueError(f'Unexpected dataset kind {kind}')
-
-    def finalize_check_result(self, check_result, check, kind: DatasetKind = None):
-        """Run final processing on a check result which includes validation, conditions processing and sampling\
-        footnote."""
-        # Validate the check result type
-        if isinstance(check_result, CheckFailure):
-            return
-        if not isinstance(check_result, CheckResult):
-            raise DeepchecksValueError(f'Check {check.name()} expected to return CheckResult but got: '
-                                       + type(check_result).__name__)
-
-        # Set reference between the check result and check
-        check_result.check = check
-        # Calculate conditions results
-        check_result.process_conditions()
-        # Add sampling footnote if needed
-        if hasattr(check, 'n_samples'):
-            n_samples = getattr(check, 'n_samples')
-            message = ''
-            if kind:
-                dataset = self.get_data_by_kind(kind)
-                if dataset.is_sampled(n_samples):
-                    message = f'Data is sampled from the original dataset, running on ' \
-                              f'{dataset.len_when_sampled(n_samples)} samples out of {len(dataset)}.'
-            else:
-                if self._train is not None and self._train.is_sampled(n_samples):
-                    message += f'Running on {self._train.len_when_sampled(n_samples)} <b>train</b> data samples ' \
-                               f'out of {len(self._train)}.'
-                if self._test is not None and self._test.is_sampled(n_samples):
-                    if message:
-                        message += ' '
-                    message += f'Running on {self._test.len_when_sampled(n_samples)} <b>test</b> data samples ' \
-                               f'out of {len(self._test)}.'
-
-            if message:
-                message = ('<p style="font-size:0.9em;line-height:1;"><i>'
-                           f'Note - data sampling: {message} Sample size can be controlled with the "n_samples" '
-                           'parameter.</i></p>')
-                check_result.display.append(message)
